requirements.txt
flask
flask-sqlalchemy
python-dotenv
uuid
### these requirements depend on the model
# torch
# transformers
# accelerate 
# sentencepiece 
# bitsandbytes 
# einops 
# tokenizers
# optimum